refactor: move purchaseCow to its own file
