+++ src/game-logic/reducers/index.js
@@ -7,7 +7,6 @@ import { itemsMap } from '../../data/maps'
 import achievements from '../../data/achievements'
 import {
   areHuggingMachinesInInventory,
-  generateCow,
   getCostOfNextStorageExpansion,
   getCowColorId,
   getCowValue,
@@ -86,37 +85,7 @@ export * from './sellItem'
 export * from './updateLearnedRecipes'
 export * from './makeRecipe'
 export * from './upgradeTool'
-
-/**
- * @param {farmhand.state} state
- * @param {farmhand.cow} cow
- * @returns {farmhand.state}
- */
-export const purchaseCow = (state, cow) => {
-  const { cowInventory, cowColorsPurchased, id, money, purchasedCowPen } = state
-  const { color } = cow
-  const cowValue = getCowValue(cow, false)
-
-  if (
-    money < cowValue ||
-    purchasedCowPen === 0 ||
-    cowInventory.length >= PURCHASEABLE_COW_PENS.get(purchasedCowPen).cows
-  ) {
-    return state
-  }
-
-  state = addCowToInventory(state, { ...cow, ownerId: id, originalOwnerId: id })
-
-  return {
-    ...state,
-    cowColorsPurchased: {
-      ...cowColorsPurchased,
-      [color]: (cowColorsPurchased[color] || 0) + 1,
-    },
-    money: moneyTotal(money, -cowValue),
-    cowForSale: generateCow(),
-  }
-}
+export * from './purchaseCow'
 
 /**
  * @param {farmhand.state} state
